clearing out demo stuff
