--- .prettier.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-  trailingComma: 'es5',
-  tabWidth: 2,
-  semi: true,
-  singleQuote: true,
-  useTabs: false,
-  jsxSingleQuote: true,
-  bracketSpacing: true,
-  bracketSameLine: true,
-};
