+++ .storybook/main.js
@@ -1,10 +1,17 @@
 module.exports = {
-  stories: ['../src/**/*.stories.@(js|jsx|ts|tsx)'],
+  stories: ['../src/components/**/*.stories.@(js|jsx|ts|tsx)'],
   addons: [
+    '@storybook/preset-create-react-app',
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/addon-interactions',
+    '@storybook/preset-scss',
+    '@storybook/preset-typescript',
   ],
+  typescript: {
+    check: false,
+    checkOptions: {},
+  },
   framework: '@storybook/react',
   core: {
     builder: '@storybook/builder-webpack5',
